refactor: useSearch훅 통해 검색 로직 분리
